# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GHIC Engineering Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /repositories:
    get:
      operationId: listRepositories
      tags: [repositories]
      summary: List all connected repositories
      parameters:
        - in: query
          name: page
          schema: { type: number }
        - in: query
          name: limit
          schema: { type: number }
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: language
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
      responses:
        "200":
          description: Repositories list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepositoryListResponse"
  /repositories/{id}:
    get:
      operationId: getRepository
      tags: [repositories]
      summary: Get repository details
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Repository details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Repository"
  /repositories/{id}/sync:
    post:
      operationId: syncRepository
      tags: [repositories]
      summary: Sync a repository
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Sync triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
  /repositories/{id}/reindex:
    post:
      operationId: reindexRepository
      tags: [repositories]
      summary: Reindex a repository
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Reindex triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
  /repositories/{id}/analytics:
    get:
      operationId: getRepositoryAnalytics
      tags: [analytics]
      summary: Get analytics for a repository
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: query
          name: period
          schema: { type: string }
      responses:
        "200":
          description: Repository analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepositoryAnalytics"
  /repositories/{id}/commits:
    get:
      operationId: getRepositoryCommits
      tags: [commits]
      summary: Get commits for a repository
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: query
          name: page
          schema: { type: number }
        - in: query
          name: limit
          schema: { type: number }
      responses:
        "200":
          description: Commits
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommitListResponse"
  /repositories/{id}/pull-requests:
    get:
      operationId: getRepositoryPullRequests
      tags: [pullrequests]
      summary: Get pull requests for a repository
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: query
          name: page
          schema: { type: number }
        - in: query
          name: status
          schema: { type: string }
      responses:
        "200":
          description: Pull requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PullRequestListResponse"
  /repositories/{id}/releases:
    get:
      operationId: getRepositoryReleases
      tags: [releases]
      summary: Get releases for a repository
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: query
          name: page
          schema: { type: number }
      responses:
        "200":
          description: Releases
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReleaseListResponse"
  /repositories/{id}/components:
    get:
      operationId: getRepositoryComponents
      tags: [components]
      summary: Get components for a repository
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: query
          name: page
          schema: { type: number }
      responses:
        "200":
          description: Components
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComponentListResponse"

  # ─── Issues ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ActionResult:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
      required: [success, message]

    # ─── Dashboard ────────────────────────────────────────────────────────────
...
    Repository:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        fullName: { type: string }
        owner: { type: string }
        visibility: { type: string }
        language: { type: ["string", "null"] }
        framework: { type: ["string", "null"] }
        defaultBranch: { type: string }
        sizeKb: { type: number }
        stars: { type: number }
        forks: { type: number }
        openIssues: { type: number }
        contributors: { type: number }
        healthScore: { type: number }
        riskScore: { type: number }
        lastIndexedAt: { type: ["string", "null"] }
        indexStatus: { type: string }
        repositoryIntelligenceStatus: { type: string }
        engineeringIntelligenceStatus: { type: string }
        webhookStatus: { type: string }
        installationStatus: { type: string }
        description: { type: ["string", "null"] }
        url: { type: string }
        createdAt: { type: string }
        updatedAt: { type: string }
      required:
        - id
        - name
        - fullName
        - owner
        - visibility
        - defaultBranch
        - sizeKb
        - stars
        - forks
        - openIssues
        - contributors
        - healthScore
        - riskScore
        - indexStatus
        - repositoryIntelligenceStatus
        - engineeringIntelligenceStatus
        - webhookStatus
        - installationStatus
        - url
        - createdAt
        - updatedAt
    RepositoryListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Repository"
        total: { type: number }
        page: { type: number }
        limit: { type: number }
      required: [items, total, page, limit]
    RepositoryAnalytics:
      type: object
      properties:
        repositoryId: { type: string }
        bugDensity: { type: number }
        issueVelocity: { type: number }
        engineeringVelocity: { type: number }
        regressionRate: { type: number }
        duplicateRate: { type: number }
        resolutionTimeDays: { type: number }
        componentStability: { type: number }
        releaseStability: { type: number }
        issuesByDay:
          type: array
          items:
            $ref: "#/components/schemas/TimeSeriesPoint"
        resolutionsByDay:
          type: array
          items:
            $ref: "#/components/schemas/TimeSeriesPoint"
        topContributors:
          type: array
          items:
            $ref: "#/components/schemas/ContributorStat"
        mostChangedFiles:
          type: array
          items:
            $ref: "#/components/schemas/FileStat"
        mostFragileComponents:
          type: array
          items:
            $ref: "#/components/schemas/ComponentStat"
      required:
        - repositoryId
        - bugDensity
        - issueVelocity
        - engineeringVelocity
        - regressionRate
        - duplicateRate
        - resolutionTimeDays
        - componentStability
        - releaseStability
    TimeSeriesPoint:
      type: object
      properties:
        date: { type: string }
        value: { type: number }
      required: [date, value]
    ContributorStat:
      type: object
      properties:
        login: { type: string }
        avatarUrl: { type: ["string", "null"] }
        commits: { type: number }
        additions: { type: number }
        deletions: { type: number }
      required: [login, commits, additions, deletions]
    FileStat:
      type: object
      properties:
        path: { type: string }
        changes: { type: number }
        lastModified: { type: string }
      required: [path, changes, lastModified]
    ComponentStat:
      type: object
      properties:
        name: { type: string }
        issueCount: { type: number }
        regressionCount: { type: number }
        healthScore: { type: number }
      required: [name, issueCount, regressionCount, healthScore]

    # ─── Issue ────────────────────────────────────────────────────────────────
...
    Commit:
      type: object
      properties:
        id: { type: string }
        sha: { type: string }
        shortSha: { type: string }
        message: { type: string }
        author: { type: string }
        authorAvatarUrl: { type: ["string", "null"] }
        repositoryId: { type: string }
        repositoryName: { type: string }
        date: { type: string }
        filesChanged: { type: number }
        additions: { type: number }
        deletions: { type: number }
        relatedIssues:
          type: array
          items: { type: string }
        relatedPrs:
          type: array
          items: { type: string }
        risk: { type: string }
        regressionSignals: { type: boolean }
        affectedComponents:
          type: array
          items: { type: string }
        aiSummary: { type: ["string", "null"] }
        url: { type: string }
      required:
        - id
        - sha
        - shortSha
        - message
        - author
        - repositoryId
        - repositoryName
        - date
        - filesChanged
        - additions
        - deletions
        - relatedIssues
        - relatedPrs
        - risk
        - regressionSignals
        - affectedComponents
        - url
    CommitListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Commit"
        total: { type: number }
        page: { type: number }
        limit: { type: number }
      required: [items, total, page, limit]

    # ─── Pull Request ─────────────────────────────────────────────────────────
    PullRequest:
      type: object
      properties:
        id: { type: string }
        number: { type: number }
        title: { type: string }
        status: { type: string }
        repositoryId: { type: string }
        repositoryName: { type: string }
        author: { type: string }
        authorAvatarUrl: { type: ["string", "null"] }
        relatedIssues:
          type: array
          items: { type: string }
        risk: { type: string }
        reviewComplexity: { type: string }
        affectedComponents:
          type: array
          items: { type: string }
        estimatedTesting: { type: string }
        linkedRelease: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
        url: { type: string }
      required:
        - id
        - number
        - title
        - status
        - repositoryId
        - repositoryName
        - author
        - relatedIssues
        - risk
        - reviewComplexity
        - affectedComponents
        - estimatedTesting
        - createdAt
        - updatedAt
        - url
    PullRequestListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/PullRequest"
        total: { type: number }
        page: { type: number }
        limit: { type: number }
      required: [items, total, page, limit]

    # ─── Release ──────────────────────────────────────────────────────────────
    Release:
      type: object
      properties:
        id: { type: string }
        version: { type: string }
        repositoryId: { type: string }
        repositoryName: { type: string }
        releaseDate: { type: string }
        issuesIntroduced: { type: number }
        issuesFixed: { type: number }
        regressionCount: { type: number }
        risk: { type: string }
        affectedComponents:
          type: array
          items: { type: string }
        health: { type: string }
        rollbackRisk: { type: string }
        url: { type: string }
      required:
        - id
        - version
        - repositoryId
        - repositoryName
        - releaseDate
        - issuesIntroduced
        - issuesFixed
        - regressionCount
        - risk
        - affectedComponents
        - health
        - rollbackRisk
        - url
    ReleaseListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Release"
        total: { type: number }
        page: { type: number }
        limit: { type: number }
      required: [items, total, page, limit]

    # ─── Component ────────────────────────────────────────────────────────────
    Component:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        repositoryId: { type: string }
        repositoryName: { type: string }
        path: { type: string }
        health: { type: string }
        healthScore: { type: number }
        risk: { type: string }
        riskScore: { type: number }
        issueCount: { type: number }
        regressionCount: { type: number }
        recentCommits: { type: number }
        recentIssues: { type: number }
        relatedFiles:
          type: array
          items: { type: string }
        mostActiveContributors:
          type: array
          items: { type: string }
        lastUpdated: { type: string }
      required:
        - id
        - name
        - repositoryId
        - repositoryName
        - path
        - health
        - healthScore
        - risk
        - riskScore
        - issueCount
        - regressionCount
        - recentCommits
        - recentIssues
        - relatedFiles
        - mostActiveContributors
        - lastUpdated
    ComponentListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Component"
        total: { type: number }
        page: { type: number }
        limit: { type: number }
      required: [items, total, page, limit]

    # ─── Duplicate ────────────────────────────────────────────────────────────
...